Add defaults and max values, added asserts, rework Timecode class clean up indents
A bunch of default and maximum values for frames in timecode and frame
count. Added assert functions and assertion for fps for all methods.
Started a new version of the Timecode class to have it focus to work
with frame count rather then the time values. Did some clean up to get
rid of indent errors